Bug #31019130 CASTING DOUBLE TO LONGLONG CAUSES UB IN FIELD_*::STORE(DOUBLE)
Conversion of double values to MySQL types bit/enum/set could generate
UBSAN warnings. Add explicit out-of-bounds checking to
Field_bit::store(double)
Field_enum::store(double)
Field_set::store(double)

Remove C-style casts when comparing doubles to integer constants like
INT_MIN16, INT_MAX32, LLONG_MAX, ULLONG_MAX. This allows clang to
generate warnings about undefined integer conversions at compile time.

+++ sql/item.h
@@ -4897,7 +4897,7 @@ class Item_float : public Item_num {
   }
   longlong val_int() override {
     DBUG_ASSERT(fixed == 1);
-    if (value <= (double)LLONG_MIN) {
+    if (value <= LLONG_MIN) {
       return LLONG_MIN;
     } else if (value > LLONG_MAX_DOUBLE) {
       return LLONG_MAX;
